requirements.txt
# Core dependencies
streamlit>=1.28.0
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
faiss-cpu>=1.7.4
rank-bm25>=0.2.2
nltk>=3.8
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0

# PDF processing
PyPDF2>=3.0.0

# Web scraping (optional - may not work on Streamlit Cloud)
# selenium>=4.15.0
# webdriver-manager>=4.0.0

# Optional: For better performance
numpy>=1.24.0
scikit-learn>=1.3.0
